Commit 1.3. Flashlight Improvements
Animations added.
Flashlight Controller.cs added.

Now the player can press the button "G" to turn on/off the flashlight.
Coroutines added as well for a short delay on the "spot light" to make it look a bit better and independent as well.

The hand now also has a simple skin color.