Commit 1.6 Spawn System
Added a spawn System where SCP-106 will be able to spawn. In more detail,
I created spawn point(s) where SCP-106 will be able  to spawn around the map.
The spawn points have a fixed Y position (for now) so the enemy can spawn no under the ground.
The Spawner takes the location, and the original prefab, and creates a clown of the prefab in that location. This Spawn System is further to be expanded.