Commit 1.7 Enemy now follows!
Added an enemy AI through navMeshAgent. The reason I used navMeshAgent is for SCP106 to be able to walk around obstacles when chasing the player such as trees, walls etc.
Spawn system adjusted. I changed the spawn system and now instead of creating new instances of the SCP106 main prefab (Clones), it teleports the main prefab to the spawn "portal". This is useful to me because I won't have multiple SCP106 enemies chasing the player. I only need one. This one will be disabled and enabled when its time for him to chase or stop chasing accordingly.
Added Enemy follow script. This script is to target the player and is attached on SCP106.
Terrain is now smaller to save my pc from running hot and its also adjusted for the scp to walk on it and not fall. Terrain is temporary until I start building the world.
Added eye orbs to the SCP106 to give the look he deserves.