Commit 1.9 Liquid Portal
Added black liquid portal for spawning. This black portal is where Scp-106 comes from when he spawns. It's animated and grows bigger until Scp-106 is gets out of it.
Portal has a countdown for disable which is the same time as the disable time for Scp-106.
For animation, I decided to alter the shader of the material in real-time where I slowly reduce the Cut off alpha in the shader to produce an animated-like action.