Commit 2.3 Corrosive trail added.
added trail (footprints).
Scp-106 is known of leaving a black corrosive substance whenever he walks or teleports around the place. I felt the need to add a footrail of his corrosive substance as a signature of his own identity.
I tried creating it using the spawner script I already had but it felt impossible at that time, so I had to use enemyFollowAI script to code and make it work.