Commit 3.6 Major terrain update
Updated terrain, trees added everywhere as well as grass. No more trees/grass will be added in the future.
Deleted some parts of the "terrain tool" package, due to some serious errors/bugs where it destroyed the whole project and could not be opened by unity editor.
Added river and ocean to the terrain.
Added more terrain to fill the gaps of the world.(This terrain is not walkable)