Commit 3.8 Death
Added a collider in front of SCP-106 that does a variety of things with the help of a script.
Added a death script.
Added some animations when the player dies
Added a canvas with images and text to show after the player dies.
Added a horror font "THE CONjURING".
Added some images of scp-106.
Added a Camera Animator.
Adjusted Scp-106 EnemyFollow Script, Scp had some problems walking to the new terrain so I had to adjust the positions of EnemyFollow script so Scp-106 could walk on the new terrain.
Adjusted "SpawnLocationSquare Script for the above same reason.

What the death script does:
when the collider touches the player,
it sets the player's movement speed to 0,
it sets the mouse sensitivity to 0,
it plays a camera death animation,
it plays the "hide flashlight" animation,
it disables the "FlashlightController" script,
it activates deathCanvas,
and stops Scp-106 from trying to chase the player.