Commit 4.0 Raycast system added
added a raycast system to be able to "pickup" the items around the world.
added a barn to the terrain
added items /objects which will be scattered around the world.
added jumpscare canvas where it works with a collider for random jumpscares on the screen.
added images for jumpscares and for the pickup items

~I tried to implement an inventory system but I decided not to, because it would take too long to create and I thought it wouldn't be so important for the few items that I added around the world.~