Commit 4.2 Ending event added
Added ending event where, if the player collects all five papers he will be prompt to go back to his vehicle where he started. Scp-106 will be there waiting for the player.
Added a separate Spawn location that is not following the player. (its close to the trigger box).
Added a triggerbox, if the player steps in some of the following things will happen:
player movement will be reduced to one.
Sensitivity is reduced to 20.
Scp-106 will be spawned.
There will be a separate death scene where it will thank the player for playing and that he has finished game.