Commit 4.3 Audio Manager added
Added audio manager to handle most of the sounds in-game but not all.
Added a play and stop method as unity already has.
Added a fade effect on audio manager to fade music or sounds where necessary.
Added more sounds.
Adjusted ending Canvas and added a basic idea of how credits would play at the end.