Commit 4.5 Main Menu Added
Added a sound when the player pick up an item
Adjusted credit scene and added the option to go back to Main menu using "Escape" key.

Main menu Added which Includes:
The main theme background, which is
Scp-106 in a breathing state,
a dark room,
and a red lighting for atmospheric purposes.
The theme represents SCP-106's pocket dimension and it serves as an Easter egg for future development purposes.
Added A Play, and quit button with functionalities and animations.
Added a soundtrack for the main menu.

Adjusted README file and sorted appropriately.