Commit 4.6 Fence walls added
Added Fences in some parts of the map so it feels richer when the player explores the corners.

Added sounds on buttons on Main Menu.
Added a pause menu Canvas with functional buttons.
Added a fade on main menu when scene is changing.
Added A script to death Canvas as well as a text to be able to go back to the Main Menu after the player die.

Adjusted one spot where a page is. I realized it was very difficult to find, so I added a liquid lake on that spot to give a more noticeable view.
Adjusted deathZone from some corrosive pool liquids.